Fix
Requirements and card view.
