+++ src/Alpaca/templates/Alpaca/index.html
@@ -29,11 +29,9 @@
             {% include "./partials/index-order-bar.html" %}
 
             <div class="al-flex-row">
-                <div class="al-flex-grow">
-                    {% for activity in activity_list %}
-                        {% include "./partials/activity-card.html" %}                
-                    {% endfor %}
-                </div>
+                {% for activity in activity_list %}
+                    {% include "./partials/activity-card.html" %}                
+                {% endfor %}
             </div>
         </div>
     </div>
