+++ src/requirements.txt
@@ -1,1 +1,2 @@
 django
+Pillow
